Drunken refactoring removing color_eyre and using anyhow + tracing to consistently and concisely handle errors
