+++ server/chroma/Cargo.toml
@@ -8,7 +8,6 @@ actix-cors = "0.6.4"
 actix-multiresponse = { version = "0.4.2", features = ["xml"] }
 actix-web = "4.3.0"
 actix-governor = "0.4.0"
-color-eyre = "0.6.2"
 proto = { version = "0.1.0", path = "../proto" }
 serde = { version = "1.0.152", features = ["derive"] }
 thiserror = "1.0.38"
@@ -32,6 +31,7 @@ cabbage = "0.1.2"
 moka = { version = "0.12.5", features = ["future"] }
 dotenv = "0.15.0"
 governor = "0.6.3"
+anyhow = "1.0.86"
 
 [dev-dependencies]
 serde_json = "1.0.93"
