+++ server/chroma/src/routes/authorization.rs
@@ -1,10 +1,10 @@
 use std::future::Future;
 use std::pin::Pin;
 
+use actix_web::{FromRequest, HttpRequest, HttpResponse, ResponseError, web};
 use actix_web::body::BoxBody;
 use actix_web::dev::Payload;
 use actix_web::http::StatusCode;
-use actix_web::{web, FromRequest, HttpRequest, HttpResponse, ResponseError};
 use tap::TapFallible;
 use thiserror::Error;
 use tracing::{info, trace};
