+++ server/chroma/src/routes/empty.rs
@@ -1,5 +1,5 @@
-use actix_web::body::BoxBody;
 use actix_web::{HttpRequest, HttpResponse, Responder};
+use actix_web::body::BoxBody;
 
 pub struct Empty;
