+++ server/chroma/src/routes/error.rs
@@ -1,6 +1,6 @@
+use actix_web::{HttpResponse, ResponseError};
 use actix_web::body::BoxBody;
 use actix_web::http::StatusCode;
-use actix_web::{HttpResponse, ResponseError};
 use thiserror::Error;
 
 pub type WebResult<T> = Result<T, Error>;
