+++ server/chroma/src/routes/redirect.rs
@@ -1,6 +1,6 @@
+use actix_web::{HttpRequest, HttpResponse, Responder};
 use actix_web::body::BoxBody;
 use actix_web::http::StatusCode;
-use actix_web::{HttpRequest, HttpResponse, Responder};
 
 pub struct Redirect(String);
