+++ server/chroma/src/routes/v1/album/list.rs
@@ -3,10 +3,10 @@ use actix_web::web;
 use futures::future::{join_all, try_join_all};
 use serde::Deserialize;
 
+use dal::DalError;
 use dal::database::{Album, Photo};
 use dal::storage_engine::aws_error::GetObjectErrorKind;
 use dal::storage_engine::error::{SdkError, StorageError};
-use dal::DalError;
 use proto::{AlbumWithCoverPhoto, ListAlbumsResponse};
 
 use crate::routes::appdata::{AlbumIdCache, WebData};
