+++ server/chroma/src/routes/v1/mod.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
-use actix_governor::governor::middleware::StateInformationMiddleware;
 use actix_governor::{Governor, GovernorConfig, GovernorConfigBuilder, PeerIpKeyExtractor};
+use actix_governor::governor::middleware::StateInformationMiddleware;
 use actix_web::web;
 use actix_web::web::ServiceConfig;
 use serde::Deserialize;
