+++ server/chroma/src/routes/v1/photo/list.rs
@@ -3,8 +3,8 @@ use actix_web::web;
 use futures::future::join_all;
 use serde::Deserialize;
 
-use dal::database::Photo;
 use dal::DalError;
+use dal::database::Photo;
 use proto::ListPhotoResponse;
 
 use crate::routes::appdata::WebData;
